requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
colorama>=0.4.6
fake-useragent>=1.4.0
selenium>=4.15.0
webdriver-manager>=4.0.0
openpyxl>=3.1.0
pyyaml>=6.0
python-dotenv>=1.0.0